feat: display config values in CLI
