+++ src-tauri/Cargo.toml
@@ -32,6 +32,7 @@ rustls-pemfile = "2.1.3"
 rodio = "0.19.0"
 clap = { version = "4.0.32", features = ["derive"] }
 tauri-plugin-dialog = "2"
+dirs = "6.0.0"
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
 # If you use cargo directly instead of tauri's cli you can use this feature flag to switch between tauri's `dev` and `build` modes.
