+++ src-tauri/src/config.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 use std::fs;
 use std::fs::OpenOptions;
 use std::io::Write;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Config {
@@ -37,12 +37,15 @@ fn default_theme() -> String {
 }
 
 impl Config {
+    pub fn get_config_file_path(config_dir: &Path, config_file_name: Option<String>) -> PathBuf {
+        config_dir.join(config_file_name.unwrap_or("config.toml".to_string()))
+    }
+
     pub fn get_or_create_from_disk(
         config_dir: &PathBuf,
         config_file_name: Option<String>,
     ) -> Result<Self, Box<dyn std::error::Error>> {
-        let file_name = config_file_name.unwrap_or("config.toml".to_string());
-        let config_file_path = config_dir.join(file_name);
+        let config_file_path = Self::get_config_file_path(config_dir, config_file_name);
 
         // Create the config dir and the themes one if they don’t exist
         let _ = fs::create_dir_all(config_dir.join("themes/"));
@@ -70,5 +73,31 @@ impl Config {
 
             config
         })
+    }
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            always_on_top: true,
+            auto_start_work_timer: true,
+            auto_start_break_timer: true,
+            desktop_notifications: true,
+            focus_audio: None,
+            focus_duration: 25 * 60,
+            long_break_audio: None,
+            long_break_duration: 20 * 60,
+            max_round_number: 4u16,
+            minimize_to_tray: true,
+            minimize_to_tray_on_close: true,
+            muted: false,
+            short_break_audio: None,
+            short_break_duration: 5 * 60,
+            start_minimized: false,
+            system_startup_auto_start: false,
+            theme: "pomotroid".to_string(),
+            tick_sounds_during_work: true,
+            tick_sounds_during_break: true,
+        }
     }
 }
